+++ src/main/java/com/example/demo/Mapper/Repository/SaveOnPostMentionRepository.java
@@ -11,5 +11,5 @@ import org.springframework.stereotype.Repository;
 public interface SaveOnPostMentionRepository extends JpaRepository<SaveOnPostMention, Long> {
     @Modifying
     @Query(value = "UPDATE afk.save_on_post_mentions SET mention_on = :status WHERE user_id = :userId", nativeQuery = true)
-    void UpdateStatus(@Param("status") Byte status, @Param("userId") Long userId);
+    void UpdateStatus(@Param("status") Integer status, @Param("userId") Long userId);
 }
